:hammer: Moving around code related to the main screen
The main screen is getting pretty large now; I don't like files that get
into the 1,000s of lines if I can help it. While this doesn't solve the
problem, it does move things around that are purely related to the main
screen, and also moves a little bit of code off the class that doesn't need
to be in the class.
