+++ src/rogallo/screens/main/text_decoder.py
